Create analytics saga
